kv: update MVCCStats to track abort span size
Previously, we didn't track the size of a replica's abort span, which
led to problems with the GC process when the abort span got too large
(#40815). The GC process was run heuristically if the sysbytes count
(which the abort span bytes count is a subset of) exceeds a certain
threshold.

Now, MVCCStats tracks abort span size, which allows us to update
the GC process to run if the actual abort span size exceeds a
threshold, rather than estimating based on the sysbytes count.

After adding the abort span bytes field to MVCC stats, the version
upgrade tests started failing. The new field caused the MVCC stats
struct to be larger than in previous versions, which resulted in the
replica consistency checker breaking due to conflicting sysbyte counts.

To address this, we add a legacy representation of MVCC stats that does
not include the abort span bytes field, and we use this representation
whenever writing to the RangeStatsLegacyKey, which resolves the above
conflicts.

Touches #50376. Closes #27055.

Release note: None